Move context to a separate file + prettier
