+++ packages/dds/tree/src/feature-libraries/editable-tree/editableTree.ts
@@ -2,8 +2,9 @@
  * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
  * Licensed under the MIT License.
  */
+
 import { assert } from "@fluidframework/common-utils";
-import { Value, Anchor, rootFieldKey } from "../../tree";
+import { Value, Anchor } from "../../tree";
 import {
     IEditableForest,
     TreeNavigationResult,
@@ -18,7 +19,6 @@ import {
     TreeSchemaIdentifier,
     TreeSchema,
     ValueSchema,
-    lookupGlobalFieldSchema,
     lookupTreeSchema,
 } from "../../schema-stored";
 import { FieldKind, Multiplicity } from "../modular-schema";
@@ -32,6 +32,7 @@ import {
     isPrimitiveValue,
     PrimitiveValue,
 } from "./utilities";
+import { EditableTreeContext, ProxyContext } from "./editableTreeContext";
 
 /**
  * A symbol for extracting target from editable-tree proxies.
@@ -153,85 +154,7 @@ export type UnwrappedEditableField =
     | undefined
     | readonly UnwrappedEditableTree[];
 
-/**
- * A common context of a "forest" of EditableTrees.
- * It handles group operations like transforming cursors into anchors for edits.
- * TODO: add test coverage.
- */
-export interface EditableTreeContext {
-    /**
-     * Gets a Javascript Proxy providing a JavaScript object like API for interacting with the tree.
-     *
-     * Use built-in JS functions to get more information about the data stored e.g.
-     * ```
-     * for (const key of Object.keys(context.root)) { ... }
-     * // OR
-     * if ("foo" in data) { ... }
-     * context.free();
-     * ```
-     *
-     * Not (yet) supported: create properties, set values and delete properties.
-     */
-    readonly root: UnwrappedEditableField;
-
-    /**
-     * Call before editing.
-     *
-     * Note that after performing edits, EditableTrees for nodes that no longer exist are invalid to use.
-     * TODO: maybe add an API to check if a specific EditableTree still exists,
-     * and only make use other than that invalid.
-     */
-    prepareForEdit(): void;
-
-    /**
-     * Call to free resources.
-     * EditableTrees created in this context are invalid to use after this.
-     */
-    free(): void;
-}
-
-class ProxyContext implements EditableTreeContext {
-    public readonly withCursors: Set<ProxyTarget> = new Set();
-    public readonly withAnchors: Set<ProxyTarget> = new Set();
-
-    constructor(public readonly forest: IEditableForest) {}
-
-    public prepareForEdit(): void {
-        for (const target of this.withCursors) {
-            target.prepareForEdit();
-        }
-        assert(this.withCursors.size === 0, 0x3c0 /* prepareForEdit should remove all cursors */);
-    }
-
-    public free(): void {
-        for (const target of this.withCursors) {
-            target.free();
-        }
-        for (const target of this.withAnchors) {
-            target.free();
-        }
-        assert(this.withCursors.size === 0, 0x3c1 /* free should remove all cursors */);
-        assert(this.withAnchors.size === 0, 0x3c2 /* free should remove all anchors */);
-    }
-
-    public get root(): UnwrappedEditableField {
-        const cursor = this.forest.allocateCursor();
-        const destination = this.forest.root(this.forest.rootField);
-        const cursorResult = this.forest.tryMoveCursorTo(destination, cursor);
-        const targets: ProxyTarget[] = [];
-        if (cursorResult === TreeNavigationResult.Ok) {
-            do {
-                targets.push(new ProxyTarget(this, cursor));
-            } while (cursor.seek(1) === TreeNavigationResult.Ok);
-        }
-        cursor.free();
-        this.forest.anchors.forget(destination);
-        const rootSchema = lookupGlobalFieldSchema(this.forest.schema, rootFieldKey);
-        return proxifyField(getFieldKind(rootSchema), targets);
-    }
-}
-
-class ProxyTarget {
+export class ProxyTarget {
     private readonly lazyCursor: ITreeSubscriptionCursor;
     private anchor?: Anchor;
 
@@ -504,7 +427,10 @@ function inProxyOrUnwrap(target: ProxyTarget): UnwrappedEditableTree {
  * @param fieldKind - determines how return value should be typed. See {@link UnwrappedEditableField}.
  * @param childTargets - targets for the children of the field.
  */
-function proxifyField(fieldKind: FieldKind, childTargets: ProxyTarget[]): UnwrappedEditableField {
+export function proxifyField(
+    fieldKind: FieldKind,
+    childTargets: ProxyTarget[],
+): UnwrappedEditableField {
     if (fieldKind.multiplicity === Multiplicity.Sequence) {
         // Return array for sequence fields
         return childTargets.map(inProxyOrUnwrap);
